feat(extension): rename to get started components LW-9403 LW-9413
