+++ packages/core/src/ui/components/SharedWallet/index.ts
@@ -1,3 +1,7 @@
+export {
+  SharedWalletGetStartedOptions,
+  SharedWalletSetupOptionTranslations
+} from './setup/SharedWalletGetStartedDialog.component';
 export { SetupSharedWallet } from './SetupSharedWallet';
 export { AllSet } from './AllSet';
 export { ImportantInfo } from './ImportantInfo';
