+++ packages/core/src/ui/components/SharedWallet/setup/SharedWalletGetStartedDialog.module.scss
