+++ packages/core/src/ui/components/SharedWallet/setup/SharedWalletGetStartedOption.module.scss
